# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для сервиса комментариев",
        "title": "commentTree API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/comments": {
            "get": {
                "description": "Получает комментарии по parentId, поддерживает фильтр search, пагинацию и сортировку",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "Get Comments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Parent ID (если не указан, можно использовать search)",
                        "name": "parent",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Текст для поиска комментариев",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Номер страницы",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Размер страницы",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "asc",
                        "description": "Сортировка asc/desc",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список комментариев с деревом вложенности",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/app.CommentNode"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid parent id",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service unavailable (DB error)",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Создает новый комментарий, можно указать ParentId для вложенного комментария",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "Create Comment",
                "parameters": [
                    {
                        "description": "Comment to create",
                        "name": "comment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.CommentReqCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created comment",
                        "schema": {
                            "$ref": "#/definitions/app.Comment"
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service unavailable (DB error)",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/comments/{id}": {
            "delete": {
                "description": "Удаляет комментарий и все его дочерние комментарии по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "Delete Comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Comment deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid comment ID",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service unavailable (DB error)",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "app.Comment": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "app.CommentNode": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app.CommentNode"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "web.CommentReqCreate": {
            "type": "object",
            "required": [
                "text"
            ],
            "properties": {
                "parent_id": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "web.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid input data"
                }
            }
        }
    }
}